ai-task-backend/requirements.txt
sqlalchemy[asyncio]
asyncpg
redis[hiredis]
openai
python-dotenv
aio-pika
